feature/AC-1 functional buttons
